feat/entity: change Dishe and Categorie in database
