+++ src/Entity/Categorie.php
@@ -1,0 +1,69 @@
+<?php
+
+namespace App\Entity;
+
+use App\Repository\CategorieRepository;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Entity(repositoryClass: CategorieRepository::class)]
+class Categorie
+{
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
+
+    #[ORM\Column(length: 255)]
+    private ?string $name = null;
+
+    #[ORM\ManyToMany(targetEntity: Dishe::class, inversedBy: 'categories')]
+    private Collection $dishe;
+
+    public function __construct()
+    {
+        $this->dishe = new ArrayCollection();
+    }
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): self
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    /**
+     * @return Collection<int, Dishe>
+     */
+    public function getDishe(): Collection
+    {
+        return $this->dishe;
+    }
+
+    public function addDishe(Dishe $dishe): self
+    {
+        if (!$this->dishe->contains($dishe)) {
+            $this->dishe->add($dishe);
+        }
+
+        return $this;
+    }
+
+    public function removeDishe(Dishe $dishe): self
+    {
+        $this->dishe->removeElement($dishe);
+
+        return $this;
+    }
+}
