--- src/Entity/Categories.php
@@ -1,71 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use App\Repository\CategoriesRepository;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\DBAL\Types\Types;
-use Doctrine\ORM\Mapping as ORM;
-
-#[ORM\Entity(repositoryClass: CategoriesRepository::class)]
-class Categories
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
-
-
-    #[ORM\ManyToMany(targetEntity: Dishes::class, inversedBy: 'categories')]
-    private Collection $dishes;
-
-    #[ORM\Column(length: 255)]
-    private ?string $name = null;
-
-    public function __construct()
-    {
-        $this->dishes = new ArrayCollection();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return Collection<int, Dishes>
-     */
-    public function getDishes(): Collection
-    {
-        return $this->dishes;
-    }
-
-    public function addDish(Dishes $dish): self
-    {
-        if (!$this->dishes->contains($dish)) {
-            $this->dishes->add($dish);
-        }
-
-        return $this;
-    }
-
-    public function removeDish(Dishes $dish): self
-    {
-        $this->dishes->removeElement($dish);
-
-        return $this;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-}
